+++ src/containerapp/azext_containerapp/_help.py
@@ -39,6 +39,22 @@ helps['containerapp create'] = """
           az containerapp create -n MyContainerapp -g MyResourceGroup \\
               --environment MyContainerappEnv \\
               --yaml "path/to/yaml/file.yml"
+    - name: Create a container app with an http scale rule
+      text: |
+          az containerapp create -n myapp -g mygroup --environment myenv --image nginx \\
+              --scale-rule-name my-http-rule \\
+              --scale-rule-http-concurrency 50
+    - name: Create a container app with a custom scale rule
+      text: |
+          az containerapp create -n MyContainerapp -g MyResourceGroup \\
+              --image my-queue-processor --environment MyContainerappEnv \\
+              --min-replicas 4 --max-replicas 8 \\
+              --scale-rule-name queue-based-autoscaling \\
+              --scale-rule-type azure-queue \\
+              --scale-rule-metadata "accountName=mystorageaccountname" \\
+                                    "cloud=AzurePublicCloud" \\
+                                    "queueLength": "5" "queueName": "foo" \\
+              --scale-rule-auth "connection=my-connection-string-secret-name"
 """
 
 helps['containerapp update'] = """
@@ -54,6 +70,18 @@ helps['containerapp update'] = """
           az containerapp update -n MyContainerapp -g MyResourceGroup \\
               --cpu 0.5 --memory 1.0Gi \\
               --min-replicas 4 --max-replicas 8
+    - name: Update a container app with an http scale rule
+      text: |
+          az containerapp update -n myapp -g mygroup \\
+              --scale-rule-name my-http-rule \\
+              --scale-rule-http-concurrency 50
+    - name: Update a container app with a custom scale rule
+      text: |
+          az containerapp update -n myapp -g mygroup \\
+              --scale-rule-name my-custom-rule \\
+              --scale-rule-type my-custom-type \\
+              --scale-rule-metadata key=value key2=value2 \\
+              --scale-rule-auth triggerparam=secretref triggerparam=secretref
 """
 
 helps['containerapp delete'] = """
@@ -323,6 +351,17 @@ helps['containerapp env create'] = """
               --location eastus2
 """
 
+helps['containerapp env update'] = """
+    type: command
+    short-summary: Update a Container Apps environment.
+    examples:
+    - name: Update an environment's custom domain configuration.
+      text: |
+          az containerapp env update -n MyContainerappEnvironment -g MyResourceGroup \\
+              --dns-suffix my-suffix.net --certificate-file MyFilePath \\
+              --certificate-password MyCertPass
+"""
+
 
 helps['containerapp env delete'] = """
     type: command
@@ -563,9 +602,9 @@ helps['containerapp ingress show'] = """
 
 helps['containerapp ingress enable'] = """
     type: command
-    short-summary: Enable ingress for a container app.
+    short-summary: Enable or update ingress for a container app.
     examples:
-    - name: Enable ingress for a container app.
+    - name: Enable or update ingress for a container app.
       text: |
           az containerapp ingress enable -n MyContainerapp -g MyResourceGroup \\
               --type external --allow-insecure --target-port 80 --transport auto